refactor: Update useHandleNodeHooks to include saveList function
This commit updates the useHandleNodeHooks hook in the useHandleNodeHooks.ts file. It adds a saveList function that allows the user to save the list of nodes and edges. The saveList function is triggered by a save button in the PanelMenu component. This change improves the functionality of the useHandleNodeHooks hook by providing a way to save the state of the nodes and edges.