refactor: Update NodeList class to adjust end node position
This commit modifies the NodeList class in the list.tsx file. It updates the position of the end node to { x: this._endNode.position.x, y: -50 } and { x: this._endNode.position.x, y: -50 } in order to adjust its vertical position. This change improves the layout of the node list by aligning the end node properly.