refactor: Add optional endDate field to NodeDataSchema
The commit modifies the NodeDataSchema in the schemas.ts file to include an optional endDate field. This field is added to the NodeDataSchema to allow for the representation of the end date of a trip or road trip. This change enhances the schema definition by providing a way to capture and validate the end date of a trip or road trip.