refactor: Adjust end node position in NodeList class
This commit modifies the NodeList class in the list.tsx file to update the position of the end node. The end node's vertical position is adjusted to { x: this._endNode.position.x, y: -50 } and { x: this._endNode.position.x, y: -50 }. This change improves the layout of the node list by aligning the end node properly.