feat: Update useHandleNodeHooks to accept tripDates as an argument
This commit updates the useHandleNodeHooks custom hook to accept the tripDates as an argument. It allows for more flexibility in setting the start and end dates of the node list. By passing the tripDates object, the hook can dynamically generate the start and end nodes with the specified dates. This enhancement improves the functionality and usability of the application.