refactor: Fix animated property in NodeList class
The code changes in this commit fix the animated property in the NodeList class. Previously, the animated property was not being used correctly, resulting in the connections between nodes not being animated. This fix ensures that the animated property is set to true, enabling the desired animation effect and enhancing the visual experience of the application.