refactor: Update NodeExtend type in createStartNodeExtend and createEndNodeExtend functions
This commit updates the NodeExtend type in the createStartNodeExtend and createEndNodeExtend functions. The type is changed from "customNode" to options.type or "tripNode" if options.type is undefined. This change improves the flexibility and configurability of the node creation process, allowing for different types of nodes to be created based on the provided options.