refactor: Add "type" property to CreateNodeOptions interface
This commit adds a "type" property to the CreateNodeOptions interface in the create-node-helpers.ts file. The "type" property can be set to either "tripNode" or "roadtrip" and allows for the creation of different types of nodes based on the provided options. This change improves the flexibility and configurability of the node creation process.