refactor: Add isChild and parentId properties to NodeData interface
This commit adds the "isChild" and "parentId" properties to the NodeData interface in the list.tsx file. These properties allow for identifying if a node is a child node and specifying its parent node. This change enhances the data structure and organization of the node hierarchy, providing better management and navigation of the nodes.