```text refactor: Add getTripDates method to NodeList class
This commit adds a new method, getTripDates, to the NodeList class in the list.tsx file. The getTripDates method returns the start and end dates of a trip based on the data stored in the start and end nodes. This enhancement improves the functionality of the NodeList class by providing a convenient way to retrieve trip dates.