```text feat: Add shareTrip function to useHandleNodeHooks
This commit adds a new function called shareTrip to the useHandleNodeHooks hook. The shareTrip function saves the current list of nodes as a trip item in the database using the setItem action from the set-item module. It returns the ID of the saved trip. This feature allows users to easily share their trips with others.