refactor: Update setItem method in RemoteDatabase class
The commit refactors the setItem method in the RemoteDatabase class in the index.ts file. It modifies the method to return a Trip object instead of an array of objects. The method now inserts a new trip into the database and returns the inserted trip as a Trip object. This enhancement improves the code by providing a more accurate representation of the inserted trip and makes it easier to work with the data in subsequent code logic.